refactor: rename and move CsvMetricExporter to infrastructure layer
the reason is writing a file is coupled with so, filesystem format and disk operations
