+++ src/main/java/io/pakland/mdas/githubstats/infrastructure/postgres/controller/PostgresMiddleware.java
@@ -1,13 +1,14 @@
 package io.pakland.mdas.githubstats.infrastructure.postgres.controller;
 
+import io.pakland.mdas.githubstats.application.MetricExporter;
 import io.pakland.mdas.githubstats.application.internal.CheckMetricsFromUserByRange;
 import io.pakland.mdas.githubstats.application.internal.GetMetricsFromUserByRange;
-import io.pakland.mdas.githubstats.domain.entity.Metric;
+import io.pakland.mdas.githubstats.domain.Metric;
 import io.pakland.mdas.githubstats.infrastructure.controller.Middleware;
-import io.pakland.mdas.githubstats.infrastructure.csv.CsvAggregationFileProcessor;
 import io.pakland.mdas.githubstats.infrastructure.shell.model.ShellRequest;
 import org.springframework.stereotype.Controller;
 
+import java.io.IOException;
 import java.util.List;
 
 @Controller
@@ -18,12 +19,14 @@ public class PostgresMiddleware extends Middleware {
 
     private final GetMetricsFromUserByRange getMetricsFromUser;
 
-    private final CsvAggregationFileProcessor csvAggregationFileProcessor;
+    private final MetricExporter metricExporter;
 
-    protected PostgresMiddleware(CheckMetricsFromUserByRange checkMetricsFromUserByRange, GetMetricsFromUserByRange getMetricsFromUser, CsvAggregationFileProcessor csvAggregationFileProcessor) {
+    protected PostgresMiddleware(CheckMetricsFromUserByRange checkMetricsFromUserByRange,
+                                 GetMetricsFromUserByRange getMetricsFromUser,
+                                 MetricExporter metricExporter) {
         this.checkMetricsFromUserByRange = checkMetricsFromUserByRange;
         this.getMetricsFromUser = getMetricsFromUser;
-        this.csvAggregationFileProcessor = csvAggregationFileProcessor;
+        this.metricExporter = metricExporter;
     }
 
     @Override
@@ -31,7 +34,13 @@ public class PostgresMiddleware extends Middleware {
 
         if (checkMetricsFromUserByRange.execute(request)) {
             List<Metric> metrics = getMetricsFromUser.execute(request);
-            csvAggregationFileProcessor.write("prueba.csv", metrics);
+
+            try {
+                metricExporter.export(metrics, "prueba.csv");
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+
             return "csv from db";
         }
